ai_content_platform/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.29
alembic==1.13.1
asyncpg==0.27.0
psycopg2===2.9.11

# Config & Validation
pydantic>=2.7,<3.0
pydantic-settings>=2.7,<3.0
python-dotenv==1.0.1

# Auth & Security
python-jose[cryptography]
python-multipart
uvicorn
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9

# HTTP client
httpx==0.27.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
aiosqlite==0.22.1
anyio==4.2.0  # For async test support (pytest warning)



# Gemini AI client
google-genai>=1.2.0


# Redis client
redis==7.1.0


# logging
python-json-logger==4.0.0
# Note: httpx DeprecationWarning for 'app' shortcut. Update code or use 'httpx[http2]' if needed.
